Create Setup
To set up a plant watering robot using an Arduino, you will need the following components:

Arduino board (Uno, Mega, etc.)
Motor shield (to control the motors)
DC water pump
Water level sensor
Moisture sensor
Relay module
Jumper wires
Power supply (battery or DC adapter)
Water container
Tubing
Plant pot
Peristaltic pump (optional)
Here are the steps to set up the plant watering robot using an Arduino:

Connect the motor shield to the Arduino board, and then connect the DC water pump to the motor shield. The motor shield will allow you to control the speed and direction of the motor.

Connect the water level sensor to the Arduino board using jumper wires. The water level sensor will help you determine when the water container is running low on water.

Connect the moisture sensor to the Arduino board using jumper wires. The moisture sensor will help you determine when the soil is dry and needs to be watered.

Connect the relay module to the Arduino board using jumper wires. The relay module will help you control the peristaltic pump or other high-powered devices.

Connect the water container to the DC water pump using tubing. The tubing should be long enough to reach the plant pot.

Connect the peristaltic pump (optional) to the relay module using jumper wires. The peristaltic pump can be used to add fertilizer to the water.

Place the moisture sensor in the plant pot, and connect it to the Arduino board using jumper wires.

Set up the code on the Arduino board to control the water pump and other devices based on the readings from the water level and moisture sensors. The code should turn on the water pump when the soil is dry and turn it off when the moisture sensor detects that the soil is moist enough. The code should also turn on the peristaltic pump (if present) when needed.

Test the system to make sure it works properly. Fill the water container, and then place the plant pot in the desired location. The plant watering robot should automatically water the plant when needed.